Modified README to specify new behavior
One big change I wanted to make to the original plugin was to have it
create a new pane if one doesn't already exist in the direction a file
is being moved.  For example, if I have the default layout, I don't want
to have to separately create a pane then move a file to it.

Also added the reset layout command and the three-pane command.