Modified carry_file_to_pane behavior
Modified the behavior of carry_file_to_pane when there is only one file
in the view being moved from, and it is the only file in the target
group.  Now it does not destroy the old group, so that if one file is
open and I move it in any direction it moves it creates the new group if
necessary and then moves the file there without destroying the now empty
first group.