Added event listener to reset layout
Added event listener to reset layout on a new window.  I didn't like
that ST2 creates a new window with the same layout of the previous
window.  This has to keep up with which views have already been
processed, though, so there's a global processed_views array that just
stores the id of the views processed.  You would have to create
thousands, maybe even tens or hundreds of thousands, of views to make
that array slow down the editor, and by then other things would have
caused a slow down.